ARC: Add new compiler define to indicate @pcl relocation support
Some old versions of binutils did not support @pcl relocations.  This
commit adds a new flag to the uClibc configuration system that detects
if the toolchain supports @pcl relocations or not.

If this relocation is supported then the define ARC_HAS_AT_PCL_RELOC
will be passed to the compiler, which is then used in the arc ldso to
choose between generating old or new style code.

This commit addresses and issue that was worked around in commit
181d410ad00cddd1d6c9f4835e129136b74c5187 (ARC: Conditionalise certain
relocations as provided by TLS tools only).  In this commit the choice
between old or new style relocations was made based on whether uClibc
was configured with native threads or not.  The problem is that a user
of a new toolchain might choose to configure without native threads.